Add getInterpreterCache wrapper to legacy IOC
